refactor: 차트 컴포넌트 네이밍 Radar 제거
