+++ src/components/Charts/components/StrengthChart/StrengthChart.module.scss
